[DEV-5] Android: emulate missing star icons with color and size
